created mean annual nan function
